Allow remapping extended keys like key_321, update README.md
Extend key codes to 10 bit (mostly uint16_t instead of uint8_t)
Improve key list printing

+++ src/evloop.cpp
@@ -5,7 +5,7 @@
 static int aux_fd = -1;
 std::vector<device> device_table;
 
-static void panic_check(uint8_t code, uint8_t pressed)
+static void panic_check(uint16_t code, uint8_t pressed)
 {
 	static uint8_t enter, backspace, escape;
 	switch (code) {
