+++ src/vkbd.h
@@ -17,5 +17,5 @@ void vkbd_mouse_move(const struct vkbd *vkbd, int x, int y);
 void vkbd_mouse_move_abs(const struct vkbd *vkbd, int x, int y);
 void vkbd_mouse_scroll(const struct vkbd *vkbd, int x, int y);
 
-void vkbd_send_key(const struct vkbd *vkbd, uint8_t code, int state);
+void vkbd_send_key(const struct vkbd *vkbd, uint16_t code, int state);
 #endif
